+++ tests/plantcv/filter/eccentricity.py
@@ -1,7 +1,7 @@
 import cv2
 import pytest
 import numpy as np
-from plantcv.plantcv.filter import eccentricity
+from plantcv.plantcv.filters import eccentricity
 
 def test_filter_eccentricity(filter_test_data):
     """Test for PlantCV."""
